refactor(tests): use maturin to build Python module
