+++ Makefile
@@ -1,15 +1,37 @@
 .PHONY: test package
 
-compile-cairo:
-	@echo "Compiling cairo files..."
-	find src -type f \( -iname "*.cairo" -and -not -iname "test_*.cairo" \) -exec cairo-compile {} --cairo_path src > build/zerosync_compiled.json \;
+BIN_DIR = ./bin
+BUILD_DIR = ./build
+ARCH = $(shell uname -p)
 
-bin/stark_parser:
-	@echo "Build parser..."
+CAIRO_PROGRAM = $(BUILD_DIR)/zerosync_compiled.json
+STARK_PARSER = $(BIN_DIR)/stark_parser
+RUST_TEST_LIB = $(BIN_DIR)/libzerosync_tests.dylib
+
+CAIRO_PROGRAM:
+	find src -type f \( -iname "*.cairo" -and -not -iname "test_*.cairo" \) \
+		-exec cairo-compile {} --cairo_path src > build/zerosync_compiled.json \;
+
+STARK_PARSER:
+	@echo "Building STARK proof parser..."
 	cd src/stark_verifier/parser; \
 	cargo build --target-dir ../../../build/parser
 	mkdir bin
 	cp build/parser/debug/parser bin/stark_parser
+
+RUST_TEST_LIB:
+	@echo "Building Rust testing library..."
+	cd tests/rust_tests; \
+	maturin develop
+	cp tests/rust_tests/target/debug/libzerosync_tests.dylib bin/libzerosync_tests.dylib
+ifeq ($(ARCH), arm)
+	# On Apple Silicon (ARM), replace the installed site-package binary with one targeting x86_64.
+	# This is required due to a lack of ARM support in Protostar.
+	cd tests/rust_tests; \
+	maturin build --target x86_64-apple-darwin
+	cp tests/rust_tests/target/x86_64-apple-darwin/debug/libzerosync_tests.dylib \
+	   $$(python -c "import site; print(site.getsitepackages()[0])")/zerosync_tests/zerosync_tests.cpython-39-darwin.so
+endif
 
 chain-proof:
 	python src/chain_proof/main.py
@@ -17,13 +39,8 @@ chain-proof:
 bridge-node:
 	python src/utreexo/bridge_node.py
 
-unit-test:
-	@echo "Running unit tests..."
-	protostar -p unit test
-
-integration-test: bin/stark_parser
-	@echo "Running integration tests..."
-	protostar -p integration test
+cairo-compile: CAIRO_PROGRAM
+	@echo "Compiling cairo files..."
 
 format-cairo:
 	@echo "Formatting cairo files..."
@@ -32,6 +49,18 @@ format-cairo:
 format-cairo-check:
 	@echo "Checking format of cairo files..."
 	cairo-format src/**/*.cairo -c
+
+rust-test-lib: RUST_TEST_LIB
+	@echo "Building Rust testing library..."
+
+unit-test:
+	@echo "Running unit tests..."
+	protostar -p unit test
+
+
+integration-test: STARK_PARSER
+	@echo "Running integration tests..."
+	protostar -p integration test
 
 clean:
 	rm -rf build
