+++ tests/rust_tests/Cargo.toml
@@ -1,9 +1,17 @@
 [package]
-name = "zerosync"
+name = "zerosync_tests"
 version = "0.1.0"
 edition = "2021"
 
-[dev-dependencies]
+[lib]
+name = "zerosync_tests"
+crate-type = ["cdylib"]
+
+[build-dependencies]
+pyo3-build-config = "0.17.3"
+
+[dependencies]
+pyo3 = { version = "0.17.3", features = ["extension-module"] }
 bincode = "1.3.3"
 serde = { version = "1.0.136", features = ["derive"] }
 serde_json = "1.0.86"
