+++ tests/rust_tests/pyproject.toml
@@ -1,0 +1,6 @@
+[build-system]
+requires = ["maturin>=0.14,<0.15"]
+build-backend = "maturin"
+
+#[tool.maturin]
+#python-source = "python"
